# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "bff_project",
    "version": "1.0.0"
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "BAD_REQUEST",
                  "UNAUTHORIZED",
                  "FORBIDDEN",
                  "NOT_FOUND",
                  "CONFLICT",
                  "PAYLOAD_TOO_LARGE",
                  "UNPROCESSABLE_ENTITY",
                  "TOO_MANY_REQUESTS",
                  "INTERNAL_ERROR",
                  "BAD_GATEWAY",
                  "SERVICE_UNAVAILABLE",
                  "GATEWAY_TIMEOUT"
                ]
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            },
            "required": [
              "code",
              "message",
              "details"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/projects/{projectId}/close": {
      "patch": {
        "tags": [
          "Projects"
        ],
        "summary": "Clôture ou suspend un projet",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Identifiant du projet",
              "example": "project-1"
            },
            "required": true,
            "description": "Identifiant du projet",
            "name": "projectId",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "done",
                      "review"
                    ],
                    "default": "done"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Projet clôturé ou suspendu",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "project": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "description": "L'identifiant unique du projet",
                          "example": "123e4567-e89b-12d3-a456-426614174000"
                        },
                        "title": {
                          "type": "string",
                          "description": "Le titre du projet",
                          "example": "Nouveau projet"
                        },
                        "description": {
                          "type": "string",
                          "description": "La description du projet",
                          "example": "Description détaillée du projet"
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "todo",
                            "in-progress",
                            "review",
                            "done"
                          ],
                          "description": "Le statut du projet",
                          "example": "todo"
                        },
                        "statusLabel": {
                          "type": "string",
                          "description": "Le libellé du statut du projet",
                          "example": "À faire"
                        },
                        "priority": {
                          "type": "string",
                          "enum": [
                            "high",
                            "medium",
                            "low"
                          ],
                          "description": "La priorité du projet",
                          "example": "high"
                        },
                        "priorityLabel": {
                          "type": "string",
                          "description": "Le libellé de la priorité du projet",
                          "example": "Haute"
                        },
                        "responsible": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "description": "L'identifiant unique de la personne",
                              "example": "123e4567-e89b-12d3-a456-426614174000"
                            },
                            "name": {
                              "type": "string",
                              "description": "Le nom de la personne",
                              "example": "John Doe"
                            },
                            "avatarUrl": {
                              "type": [
                                "string",
                                "null"
                              ],
                              "format": "uri",
                              "description": "L'URL de l'avatar de la personne",
                              "example": "https://example.com/avatar.jpg"
                            }
                          },
                          "required": [
                            "id",
                            "name",
                            "avatarUrl"
                          ],
                          "description": "Une personne assignable à un projet"
                        },
                        "assignees": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "description": "L'identifiant unique de la personne",
                                "example": "123e4567-e89b-12d3-a456-426614174000"
                              },
                              "name": {
                                "type": "string",
                                "description": "Le nom de la personne",
                                "example": "John Doe"
                              },
                              "avatarUrl": {
                                "type": [
                                  "string",
                                  "null"
                                ],
                                "format": "uri",
                                "description": "L'URL de l'avatar de la personne",
                                "example": "https://example.com/avatar.jpg"
                              }
                            },
                            "required": [
                              "id",
                              "name",
                              "avatarUrl"
                            ],
                            "description": "Une personne assignable à un projet"
                          },
                          "description": "La liste des personnes assignées au projet"
                        },
                        "labels": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "description": "La liste des étiquettes associées au projet",
                          "example": [
                            "frontend",
                            "urgent"
                          ]
                        },
                        "progress": {
                          "type": "number",
                          "minimum": 0,
                          "maximum": 100,
                          "description": "Le pourcentage d'avancement du projet",
                          "example": 75
                        },
                        "dueDate": {
                          "type": "string",
                          "description": "La date d'échéance du projet au format ISO 8601",
                          "example": "2024-12-31T23:59:59Z"
                        },
                        "createdAt": {
                          "type": "string",
                          "description": "La date de création du projet au format ISO 8601",
                          "example": "2024-01-01T12:00:00Z"
                        },
                        "tasks": {
                          "type": "object",
                          "properties": {
                            "total": {
                              "type": "number",
                              "description": "Le nombre total de tâches dans le projet",
                              "example": 10
                            },
                            "completed": {
                              "type": "number",
                              "description": "Le nombre de tâches complétées dans le projet",
                              "example": 7
                            }
                          },
                          "required": [
                            "total",
                            "completed"
                          ],
                          "description": "Les statistiques des tâches du projet"
                        },
                        "permissions": {
                          "type": "object",
                          "properties": {
                            "canView": {
                              "type": "boolean",
                              "description": "Indique si l'utilisateur peut voir le projet",
                              "example": true
                            },
                            "canEdit": {
                              "type": "boolean",
                              "description": "Indique si l'utilisateur peut éditer le projet",
                              "example": true
                            },
                            "canDuplicate": {
                              "type": "boolean",
                              "description": "Indique si l'utilisateur peut dupliquer le projet",
                              "example": false
                            },
                            "canDelete": {
                              "type": "boolean",
                              "description": "Indique si l'utilisateur peut supprimer le projet",
                              "example": false
                            },
                            "canCreateTask": {
                              "type": "boolean",
                              "description": "Indique si l'utilisateur peut créer des tâches dans le projet",
                              "example": true
                            },
                            "canAssignMembers": {
                              "type": "boolean"
                            },
                            "canClose": {
                              "type": "boolean"
                            }
                          },
                          "required": [
                            "canView",
                            "canEdit",
                            "canDuplicate",
                            "canDelete",
                            "canCreateTask",
                            "canAssignMembers",
                            "canClose"
                          ],
                          "description": "Les permissions de l'utilisateur sur le projet"
                        }
                      },
                      "required": [
                        "id",
                        "title",
                        "description",
                        "status",
                        "statusLabel",
                        "priority",
                        "priorityLabel",
                        "responsible",
                        "assignees",
                        "labels",
                        "progress",
                        "dueDate",
                        "createdAt",
                        "tasks",
                        "permissions"
                      ],
                      "description": "Un projet pour les vues de type carte, table ou kanban"
                    },
                    "taskItems": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "description": "L'identifiant unique de la tâche",
                            "example": "123e4567-e89b-12d3-a456-426614174000"
                          },
                          "title": {
                            "type": "string",
                            "description": "Le titre de la tâche",
                            "example": "Nouvelle tâche"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "todo",
                              "in-progress",
                              "review",
                              "done"
                            ],
                            "description": "Le statut du projet",
                            "example": "todo"
                          },
                          "statusLabel": {
                            "type": "string",
                            "description": "Le libellé du statut de la tâche",
                            "example": "À faire"
                          },
                          "responsible": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "description": "L'identifiant unique de la personne",
                                "example": "123e4567-e89b-12d3-a456-426614174000"
                              },
                              "name": {
                                "type": "string",
                                "description": "Le nom de la personne",
                                "example": "John Doe"
                              },
                              "avatarUrl": {
                                "type": [
                                  "string",
                                  "null"
                                ],
                                "format": "uri",
                                "description": "L'URL de l'avatar de la personne",
                                "example": "https://example.com/avatar.jpg"
                              }
                            },
                            "required": [
                              "id",
                              "name",
                              "avatarUrl"
                            ],
                            "description": "Une personne assignable à un projet"
                          },
                          "assignees": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "id": {
                                  "type": "string",
                                  "description": "L'identifiant unique de la personne",
                                  "example": "123e4567-e89b-12d3-a456-426614174000"
                                },
                                "name": {
                                  "type": "string",
                                  "description": "Le nom de la personne",
                                  "example": "John Doe"
                                },
                                "avatarUrl": {
                                  "type": [
                                    "string",
                                    "null"
                                  ],
                                  "format": "uri",
                                  "description": "L'URL de l'avatar de la personne",
                                  "example": "https://example.com/avatar.jpg"
                                }
                              },
                              "required": [
                                "id",
                                "name",
                                "avatarUrl"
                              ],
                              "description": "Une personne assignable à un projet"
                            },
                            "description": "La liste des personnes assignées à la tâche"
                          },
                          "priority": {
                            "type": "string",
                            "enum": [
                              "high",
                              "medium",
                              "low"
                            ],
                            "description": "La priorité du projet",
                            "example": "high"
                          },
                          "priorityLabel": {
                            "type": "string",
                            "description": "Le libellé de la priorité de la tâche",
                            "example": "Moyenne"
                          },
                          "labels": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            },
                            "description": "La liste des étiquettes associées à la  tâche",
                            "example": [
                              "backend",
                              "important"
                            ]
                          },
                          "dueDate": {
                            "type": "string",
                            "description": "La date d'échéance de la tâche au format ISO 8601",
                            "example": "2024-12-31T23:59:59Z"
                          },
                          "completed": {
                            "type": "boolean",
                            "description": "Indique si la tâche est complétée",
                            "example": false
                          },
                          "createdAt": {
                            "type": "string",
                            "description": "La date de création de la tâche au format ISO 8601",
                            "example": "2024-01-01T12:00:00Z"
                          },
                          "updatedAt": {
                            "type": "string",
                            "description": "La date de dernière mise à jour de la tâche au format ISO 8601",
                            "example": "2024-01-15T15:30:00Z"
                          },
                          "permissions": {
                            "type": "object",
                            "properties": {
                              "canView": {
                                "type": "boolean"
                              },
                              "canEdit": {
                                "type": "boolean"
                              },
                              "canDelete": {
                                "type": "boolean"
                              },
                              "canUpdateStatus": {
                                "type": "boolean"
                              },
                              "canComment": {
                                "type": "boolean"
                              }
                            },
                            "required": [
                              "canView",
                              "canEdit",
                              "canDelete",
                              "canUpdateStatus",
                              "canComment"
                            ]
                          }
                        },
                        "required": [
                          "id",
                          "title",
                          "status",
                          "statusLabel",
                          "responsible",
                          "assignees",
                          "priority",
                          "priorityLabel",
                          "labels",
                          "dueDate",
                          "completed",
                          "createdAt",
                          "permissions"
                        ],
                        "description": "Une tâche associée à un projet"
                      }
                    }
                  },
                  "required": [
                    "project",
                    "taskItems"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient rights",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Project or task not found or not visible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "BFF User, Project API or Core API is unreachable or failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}